refactor: introducing proper layers
