+++ server/internal/exec/server/handler/ping.go
